agents/voice-segmenter/requirements.txt
# LiveKit Core 
livekit 
livekit-agents 
livekit-plugins-silero 
 
# Audio Processing 
numpy 
scipy 
soundfile 
 
# AI/Translation 
google-generativeai 
 
# Utilities 
python-dotenv 
aiofiles
